Added a layered structure to Neurons and implements feedback loop for each tick. Needs profiling

+++ Neurotics/src/world/scent/IScented.java
@@ -6,7 +6,7 @@ public interface IScented {
 	
 	public Vecf getPosition();
 	
-	public byte smell();
+	public short smell();
 	
 	public Scent getScent();
 }
